+++ Riot/Modules/Room/DataSources/RoomDataSource.h
@@ -99,4 +99,7 @@
 
 - (void)roomDataSource:(RoomDataSource*)roomDataSource didUpdateEncryptionTrustLevel:(RoomEncryptionTrustLevel)roomEncryptionTrustLevel;
 
+- (void)roomDataSource:(RoomDataSource*)roomDataSource didCancel:(MXEvent *)event;
+
+
 @end